vector: Add ShumateVectorSprite
Make ShumateVectorSpriteSheet public and add ShumateVectorSprite. This
new API gives applications more control over the sprites in a style. It
enables HiDPI support and allows apps to generate their own sprites on
demand using a callback.

+++ shumate/vector/shumate-vector-collision.c
@@ -607,7 +607,7 @@ shumate_vector_collision_visualize (ShumateVectorCollision *self,
   GHashTableIter bucket_rows;
   RTreeBucketRow *bucket_row;
 
-  double width[4] = { 1, 1, 1, 1 };
+  float width[4] = { 1, 1, 1, 1 };
   GdkRGBA color[4], color2[4];
 
   gdk_rgba_parse (&color[0], "#FF0000");
